--- a/dev-resources/logback-test.xml
+++ b/dev-resources/logback-test.xml
@@ -8,9 +8,11 @@
 
   <appender name="STDOUT_JSON" class="ch.qos.logback.core.ConsoleAppender">
     <encoder class="net.logstash.logback.encoder.LogstashEncoder">
+      <!-- <includeCallerData>true</includeCallerData> -->
       <fieldNames>
         <timestamp>timestamp</timestamp>
         <version>[ignore]</version>
+        <levelValue>[ignore]</levelValue>
       </fieldNames>
     </encoder>
   </appender>