before changes:

```
 #'mokujin.log-bench/mokujin-log-all : 1.763753 µs
 #'mokujin.log-bench/tools-logging-log-all : 118.916261 ns
 #'mokujin.log-bench/tools-logging+context : 749.481706 ns
```

Latest benchmark:

```
 #'mokujin.log-bench/mokujin-log-all : 787.695659 ns
 #'mokujin.log-bench/tools-logging-log-all : 121.793190 ns
 #'mokujin.log-bench/tools-logging+context : 445.952752 ns

```